refactor: address review nitpicks
- Return Duration.ZERO from InstantProbeAwaitBehavior.getAssertionInterval()
  since instant probing has no polling interval
- Rename ConditionalScenarioApiIT → ConditionalBehaviorApiIT
- Rename ConditionalScenarioBeforeEachIT → ConditionalBehaviorBeforeEachIT
