+++ testing/camunda-process-test-java/src/main/java/io/camunda/process/test/impl/assertions/util/InstantProbeAwaitBehavior.java
@@ -15,7 +15,6 @@
  */
 package io.camunda.process.test.impl.assertions.util;
 
-import io.camunda.process.test.api.CamundaAssert;
 import io.camunda.process.test.api.CamundaAssertAwaitBehavior;
 import java.time.Duration;
 import org.assertj.core.api.SoftAssertionsProvider.ThrowingRunnable;
@@ -54,6 +53,6 @@ public class InstantProbeAwaitBehavior implements CamundaAssertAwaitBehavior {
 
   @Override
   public Duration getAssertionInterval() {
-    return CamundaAssert.DEFAULT_ASSERTION_INTERVAL;
+    return Duration.ZERO;
   }
 }
